backend/scripts/requirements.txt
pymongo==4.7.0
Faker==18.9.0
python-dateutil==2.8.2
